fix(sc-lint): satisfy qa-10 graph follow-ups
