+++ crates/sc-lint-boundary/src/graph.rs
@@ -3,6 +3,10 @@ use crate::directive_parser::AttributeInput;
 use crate::directive_parser::Directive;
 use crate::render::hex_encode;
 
+mod reference_collector;
+
+use self::reference_collector::collect_references_with;
+
 fn collect_owner_names(items: &[Item]) -> BTreeSet<String> {
     let mut names = BTreeSet::new();
     for item in items {
@@ -69,90 +73,6 @@ impl NodeKind {
             Self::TraitRef => "trait_ref",
         }
     }
-}
-
-#[derive(Default)]
-struct ReferenceCollector {
-    owner_name: Option<String>,
-    local_owner_names: BTreeSet<String>,
-    references: BTreeSet<CollectedReference>,
-}
-
-impl ReferenceCollector {
-    fn new(local_owner_names: &BTreeSet<String>, owner_name: Option<&str>) -> Self {
-        Self {
-            owner_name: owner_name.map(ToOwned::to_owned),
-            local_owner_names: local_owner_names.clone(),
-            references: BTreeSet::new(),
-        }
-    }
-
-    fn into_references(self) -> BTreeSet<CollectedReference> {
-        self.references
-    }
-
-    fn maybe_insert_path(&mut self, path: &syn::Path, kind: ReferenceKind) {
-        let mut segments: Vec<String> = path
-            .segments
-            .iter()
-            .map(|segment| segment.ident.to_string())
-            .collect();
-        if segments.is_empty() {
-            return;
-        }
-
-        if segments[0] == "Self" {
-            if let Some(owner_name) = &self.owner_name {
-                segments[0] = owner_name.clone();
-                self.references.insert(CollectedReference {
-                    path: segments.join("::"),
-                    kind,
-                });
-            }
-            return;
-        }
-
-        if matches!(kind, ReferenceKind::Expr) && segments.len() == 1 && segments[0] == "self" {
-            return;
-        }
-
-        let first = &segments[0];
-        let last = segments.last().unwrap();
-        let should_collect = matches!(first.as_str(), "crate" | "self" | "super")
-            || self.local_owner_names.contains(first)
-            || self.local_owner_names.contains(last);
-
-        if should_collect {
-            self.references.insert(CollectedReference {
-                path: segments.join("::"),
-                kind,
-            });
-        }
-    }
-}
-
-impl<'ast> Visit<'ast> for ReferenceCollector {
-    fn visit_type_path(&mut self, type_path: &'ast syn::TypePath) {
-        self.maybe_insert_path(&type_path.path, ReferenceKind::Type);
-        syn::visit::visit_type_path(self, type_path);
-    }
-
-    fn visit_expr_path(&mut self, expr_path: &'ast syn::ExprPath) {
-        self.maybe_insert_path(&expr_path.path, ReferenceKind::Expr);
-        syn::visit::visit_expr_path(self, expr_path);
-    }
-
-    fn visit_receiver(&mut self, _receiver: &'ast Receiver) {}
-}
-
-fn collect_references_with(
-    local_owner_names: &BTreeSet<String>,
-    owner_name: Option<&str>,
-    visit: impl FnOnce(&mut ReferenceCollector),
-) -> BTreeSet<CollectedReference> {
-    let mut collector = ReferenceCollector::new(local_owner_names, owner_name);
-    visit(&mut collector);
-    collector.into_references()
 }
 
 pub(crate) fn build_workspace_graph(root: &Path) -> Result<GraphExport> {
@@ -285,6 +205,8 @@ fn ingest_module_items(
                                 format!("while resolving module `{child_module_path}`")
                             })?;
                     let child_module_dir = if has_explicit_module_path(&item_mod.attrs) {
+                        // #[path = "..."] points at the real child source file, so the child
+                        // module directory must be derived from that resolved location.
                         child_source_path
                             .parent()
                             .map(Path::to_path_buf)
@@ -292,6 +214,8 @@ fn ingest_module_items(
                     } else if child_source_path.file_name().and_then(|name| name.to_str())
                         == Some("mod.rs")
                     {
+                        // mod.rs is itself the child module root, so sibling lookups should
+                        // start from the file's parent directory.
                         child_source_path
                             .parent()
                             .map(Path::to_path_buf)
@@ -566,9 +490,9 @@ fn ingest_module_items(
                     source_path: Some(source_path.display().to_string()),
                     module_path: Some(module_path.to_string()),
                     impl_kind: Some(if trait_path.is_some() {
-                        "trait"
+                        ImplKind::Trait
                     } else {
-                        "inherent"
+                        ImplKind::Inherent
                     }),
                     impl_trait: trait_path.clone(),
                     attributes: Vec::new(),
@@ -606,9 +530,9 @@ fn ingest_module_items(
                             source_path: Some(source_path.display().to_string()),
                             module_path: Some(module_path.to_string()),
                             impl_kind: Some(if item_impl.trait_.is_some() {
-                                "trait"
+                                ImplKind::Trait
                             } else {
-                                "inherent"
+                                ImplKind::Inherent
                             }),
                             impl_trait: trait_path.clone(),
                             attributes: parse_lint_attributes(&method.attrs)?,
@@ -966,6 +890,7 @@ fn parse_rust_file(path: &Path) -> Result<File> {
 fn impl_owner_name(self_ty: &Type) -> Result<String> {
     match self_ty {
         Type::Path(type_path) => {
+            // `syn::Type::Path` always stores at least one segment for a valid path type.
             if let Some(segment) = type_path.path.segments.last() {
                 Ok(segment.ident.to_string())
             } else {
